+++ packages/connect/deno/shim.node.ts
@@ -1,4 +1,6 @@
 // See https://github.com/fromdeno/deno2node#shimming
 
-export { fetch, File, FormData, Headers, Request, Response } from "undici";
+export { default as fetch, Headers, Request, Response } from "node-fetch";
+export { default as FormData } from "form-data";
+export { File } from "@web-std/file";
 export { Deno } from "@deno/shim-deno";
