+++ packages/connect/deno/types.storage.node.ts
@@ -1,0 +1,10 @@
+import { Result } from "./types.ts";
+
+export interface HyperStorage {
+  upload: (
+    name: string,
+    data: string | Uint8Array,
+  ) => Promise<Result>;
+  download: (name: string) => Promise<NodeJS.ReadableStream>;
+  remove: (name: string) => Promise<Result>;
+}
