+++ packages/connect/node/harness/cjs/package.json
@@ -3,7 +3,7 @@
   "type": "commonjs",
   "license": "Apache-2.0",
   "scripts": {
-    "start": "node index.js",
+    "start": "node --enable-source-maps index.js",
     "preinstall": "cd ../../../ && yarn link",
     "postinstall": "yarn link hyper-connect"
   },
