+++ packages/connect/package.json
@@ -23,15 +23,18 @@
   },
   "dependencies": {
     "@deno/shim-deno": "^0.5.0",
+    "@web-std/file": "^3.0.2",
     "crocks": "^0.12.4",
-    "jose": "^4.8.0",
+    "form-data": "^4.0.0",
+    "jose": "^4.8.1",
     "ms": "^2.1.3",
-    "ramda": "^0.28.0",
-    "undici": "^5.1.1"
+    "node-fetch": "^2.6.7",
+    "ramda": "^0.28.0"
   },
   "devDependencies": {
     "@types/ms": "^0.7.31",
-    "@types/node": "^17.0.32",
+    "@types/node": "^17.0.33",
+    "@types/node-fetch": "^2.6.1",
     "@types/ramda": "^0.28.12",
     "microbundle": "^0.15.0"
   }
